add delete dialog

+++ app/src/main/java/com/apps/expensesapp/Debtor.java
@@ -4,11 +4,11 @@ public class Debtor {
 
     public final int id;
     public final String name;
-    public final double userDebt;
-    public final double gap;
-    public final double debt;
+    public final int userDebt;
+    public final int gap;
+    public final int debt;
 
-    public Debtor(int id, String name, double userDebt, double debt) {
+    public Debtor(int id, String name, int userDebt, int debt) {
         this.id = id;
         this.name = name;
         this.userDebt = userDebt;
